pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-downscaler"
version = "0.1.0"
authors = [
  { name="Sam Lader", email="me@lader.io" },
]
description = "A tool to scale down AWS resources during non-work hours"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.26.0",
    "click>=8.0.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
    "pyyaml>=6.0.0",
    "schedule>=1.2.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "moto>=4.2.0",
    "freezegun>=1.2.0",
]
dev = [
    "mypy>=1.7.0",
    "flake8>=6.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "types-boto3>=1.0.2",
    "types-pytz>=2023.3.1.1",
    "types-python-dateutil>=2.8.19.14",
    "types-PyYAML>=6.0.12.12",
]

[project.scripts]
aws-downscaler = "aws_downscaler.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["aws_downscaler*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=aws_downscaler --cov-report=term-missing"

[tool.coverage.run]
source = ["aws_downscaler"]
omit = ["tests/*"]

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503", "E501"]
per-file-ignores = ["__init__.py:F401"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]
